Create Soley-Studio.md as a OData Producer
OData service has been lately integrated into the Soley Studio tool primary for distribution of analyzed data to other business intelligence tools that companies are using to visualize the data and present it within systematic and automated reporting system. We kindly request to include the Soley Studio to the list of OData producers.